leaving more fields for Data Mining, like TrafficDirection, RW_Type, Snow_Priority, etc, in the SQL query.
The transformed file increased from 100 MB to 144 MB with the added
fields in the SQL query. The New York City Department of City Planning's
LION Single-line Street GeoDB has many fields, which adding them can
be interesting, like:

   LAtomicPolygon, RAtomicPolygon: Minimal Undivided Polygons of extremes

   NodeLevelIF, NodeLevelIT: Altitude of the extreme points in reference to
                             the shoreline/water-level.

   LZip, RZip: (Zip codes of extreme points of segments)

   NYPDID:  the New York City Police Department Area ID to which this street
            segment belongs

   L_CD, R_CD: (Community District and Borough of extreme points)

   L*2010, R*2010: Several fields which refer to Census Data from 2010
                   (similarly from field L*2000, R*2000 from year 2000,
                    and L*1990, R*1990 from year 1990 with changes in
                    the data around this same Street Segment from those
                    three different Census years)

with different meanings among fields (e.g., between the fields "SegmentID"
and "LIONKey")

All these fields in the NYC LION GeoDB can be added in the future (but the
size of the resulting GeoDB we would have to analyze in real-time would
increase too).